Update Java code to remove unchecked or unsafe operations.
ReG_SteerParameter has new public contructors to complement the create factory
methods. Constructors and factory methods now all require an initial value so
that the proper typing can be used. A slight relaxation of this policy remains
the factory method used by the JNI layer which uses a String to pass the initial
value and a type marker (REG_INT, etc) to create the correct type of parameter.

FilenameFilter now uses a properly typed Hashtable.

The MiniApp and Sink example applications are updated in line with the above
changes.